优化数据模型，设计QuestionDto用来接收前端的数据

+++ cx-bbs-model/src/main/java/com/cx/model/article/pojos/Question.java
@@ -38,7 +38,7 @@ public class Question implements Serializable {
     private String title;
 
     @ApiModelProperty("文章作者的ID")
-    private Integer authorId;
+    private Long authorId;
 
     @ApiModelProperty("点赞数量")
     private Integer likes;
